allow different type for output signal buffer

+++ extern/corsika/include/C8SignalCalculator.hh
@@ -14,7 +14,7 @@ public:
   void get_antenna_location(std::array<float, 3>& ant_xyz);
     
   void calculate(const std::array<float, 4>& track_start_xyzt, const std::array<float, 4>& track_end_xyzt, float track_charge,
-		 float t_sig_start, float t_sig_samp, std::size_t num_samples, std::vector<float>& signal, float weight = 1.0);
+		 float t_sig_start, float t_sig_samp, std::size_t num_samples, std::vector<double>& signal, float weight = 1.0);
   
 private:
